% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducible-cache.R
\name{addin_clearCache}
\alias{addin_clearCache}
\title{RStudio addin to clear recent cache entries}
\usage{
addin_clearCache()
}
\description{
RStudio addin to clear recent cache entries
}
\author{
Alex Chubaty
}
