% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-all.R
\name{addin_loadAll}
\alias{addin_loadAll}
\title{Quickly reload a package from source}
\usage{
addin_loadAll(path = getOption("spades.loadpkgs", "~/GitHub/SpaDES.core"))
}
\arguments{
\item{path}{Path to the source package directory.
Default: \code{getOption("spades.loadpkgs", "~/GitHub/SpaDES.core")}.}
}
\description{
Simply a convenience wrapper around \code{devtools::loadall}, to be used for
binding an RStudio keyboard shortcut.
By default, this will load the \pkg{SpaDES.core} package from the path
\file{~/GitHub/SpaDES.core}.
This can be changed by setting the \code{spades.loadpkgs} option.
}
\author{
Alex Chubaty
}
