% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-modules.R
\name{addinCopyModule}
\alias{addinCopyModule}
\title{RStudio addin to copy an existing module}
\usage{
addinCopyModule()
}
\description{
RStudio addin to copy an existing module
}
\author{
Alex Chubaty
}
