% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-addins-package.R
\docType{package}
\name{SpaDES.addins-package}
\alias{SpaDES.addins}
\alias{SpaDES.addins-package}
\title{The \code{SpaDES.addins} package}
\description{
Provides 'RStudio' addins for 'SpaDES' packages and 'SpaDES' module development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://predictiveecology.github.io/SpaDES.addins/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.addins}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.addins/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alexander.chubaty@canada.ca}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
