% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-addins-package.R
\docType{package}
\name{SpaDES.addins-package}
\alias{SpaDES.addins}
\alias{SpaDES.addins-package}
\title{The \code{SpaDES.addins} package}
\description{
Provides 'RStudio' addins for 'SpaDES' packages and 'SpaDES' module development.
}
\section{\code{SpaDES} module development}{


\strong{RStudio addins:}

\itemize{
  \item copy an existing \code{SpaDES} module
  \item create a new \code{SpaDES} module
}

\strong{helper functions:}

\itemize{
  \item \emph{none yet}
}
}

\section{\code{SpaDES} package development}{


\strong{RStudio addins:}

\itemize{
  \item load the development version of \pkg{SpaDES} from \file{~/GitHub/SpaDES}.
}

\strong{helper functions:}

\itemize{
  \item \emph{none yet}
}
}

\section{Package options}{


#' \code{SpaDES} packages use the following \code{\link{options}} to configure behaviour:

\itemize{
  \item \code{spades.loadpkgs}: The default packages to load using the
  \emph{SpaDES: load all} addin. Should be a path to a package's source directory.
  Default is \code{"~/GitHub/SpaDES.core"}.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://predictiveecology.github.io/SpaDES.addins/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.addins}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.addins/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alexander.chubaty@canada.ca}

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
