% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spades-addins-package.R
\docType{package}
\name{SpaDES.addins-package}
\alias{SpaDES.addins}
\alias{SpaDES.addins-package}
\title{The \code{SpaDES.addins} package}
\description{
Provides 'RStudio' addins for 'SpaDES' packages and 'SpaDES' module development.
}
\section{\code{SpaDES} module development}{


\strong{RStudio addins:}

\itemize{
  \item copy an existing \code{SpaDES} module
  \item create a new \code{SpaDES} module
}

\strong{helper functions:}

\itemize{
  \item \emph{none yet}
}
}

\section{\code{SpaDES} package development}{


\strong{RStudio addins:}

\itemize{
  \item load the development version of \pkg{SpaDES} from \file{~/GitHub/SpaDES}.
}

\strong{helper functions:}

\itemize{
  \item \emph{none yet}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://spades-addins.predictiveecology.org/}
  \item \url{https://github.com/PredictiveEcology/SpaDES.addins}
  \item Report bugs at \url{https://github.com/PredictiveEcology/SpaDES.addins/issues}
}

}
\author{
\strong{Maintainer}: Alex M Chubaty \email{alex.chubaty@gmail.com} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

Other contributors:
\itemize{
  \item Her Majesty the Queen in Right of Canada, as represented by the Minister of Natural Resources Canada [copyright holder]
}

}
